A very strange error with newlines and comments.
Code like this would fail:

    function a takes nothing returns nothing
    //\x0d
    endfunction

The '\x0d' being the literal character.
Before this commit comments were parsed like this:

    "//".*{NEWLINE}

As the flex doc stats that '.' only parses non-newline bytes.
But for some reason the '.*' above slurped past the newline which
ofcourse caused all kinds of havoc.

So the trick was to simply state that i don't want any newlines.