Added checks for natives which behave bad in a globals block.
Added a bunch of natives which either return null or crash the game
when called in a globals block.
I added a new error-type called `runtimeerror` which can be disabled
via `+noruntimeerror`. It is intended to signal errors which don't prevent
the game from loading but which behave bad once the game is running.
Might move some errors to this category later.