Added check for bad StringHash usage
If you pass `+checkstringhash` to pjass, it will now look at calls to
`StringHash` with a literal string as its parameter and store those
strings with its hash. If a string is found which hashes to the same hash
but is different to the one stored previously we report that as an error.

This usage could be useful for those who use hashtables as a kind of
struct where a hash collision could lead to some very hard to find bugs.

But since this takes more memory and is not very exact we don't enable
this by default.