Update annotations before processing function
This error came about as i wanted to silence a multiple defined function,
which currently is defined as 'semanticerror' and putting a +nosemanticerror
before the offending function didn't work. Turns out we put the function
into the hashtable (and check its name) before we update the flags. So this
simply fixes the order.