Redone scoping rules.
Somewhere in the recent patches the scoping of variables has changed. This is
developed against the latest patch as of writing (1.36.1) to reflect these new
rules. This actually makes some prior fail cases working and some working cases
failing now.

While at it i also tried to de-tangle some code and split it up more naturally.
Of course this brings some larger changes but we rely on the test-cases to
catch them. Also this is WIP so we have revisit the test-cases anyways.